AvetiumAPI/avetium/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
